requirements.txt
# API
PyJWT==2.8.0
Flask==2.0.3
Werkzeug==2.2.2

# BILLING
### KILLBILL
./conser/modules/clients/billing/killbill/killbill_fork
#git+https://github.com/alces-flight/killbill_fork.git

# CLOUD
### OPENSTACK
keystoneauth1==4.5.0
python-heatclient==2.5.1
python-novaclient==17.7.0
python-keystoneclient==4.4.0
python-cloudkittyclient==4.7.0
python-cinderclient==8.3.0
python-ceilometerclient==2.9.0
gnocchiclient==7.0.7
pika==1.3.1
